Move code to read COSMO-D2 forecasts as timeseries to separate file
