+++ lib/forecast_utils.js
@@ -8,35 +8,18 @@ const moment = require('moment')
 const path = require('path')
 const _ = require('lodash')
 const gu = require('../lib/general_utils.js')
-const su = require('../lib/stations_utils.js')
-const gf = require('../lib/grib_functions')
 const bunyan = require('bunyan')
 const log = bunyan.createLogger({
   name: 'dwd-csv-handler',
   serializers: bunyan.stdSerializers
 })
 
+const cd2u = require('./cosmo-d2_utils')
+
 const secInHour = 3600
 const milliseconds = 1000
 const hoursInDay = 24
 
-const cosmoD2AvailableFrom = moment.utc('2018051509', 'YYYYMMDDHH')
-
-function getGribDirectory (referenceTimestamp, WEATHER_DATA_BASE_PATH) {
-  let gribBaseDirectory = null
-  if (moment.utc(referenceTimestamp).isBefore(cosmoD2AvailableFrom)) {
-    gribBaseDirectory = path.join(
-      WEATHER_DATA_BASE_PATH,
-      'weather',
-      'cosmo',
-      'de',
-      'grib'
-    )
-  } else {
-    gribBaseDirectory = path.join(WEATHER_DATA_BASE_PATH, 'weather', 'cosmo-d2', 'grib')
-  }
-  return gribBaseDirectory
-}
 
 function createDescriptionString (
   vois,
@@ -218,7 +201,7 @@ function getConfigFromModel (model, COSMO_DATA_BASE_PATH, MOSMIX_DATA_BASE_PATH)
     return {
       model: 'cosmo-d2',
       allowedModelRuns: ['00', '03', '06', '09', '12', '15', '18', '21'],
-      functionToReadData: readDataCosmoD2,
+      functionToReadData: cd2u.readTimeseriesDataCosmoD2,
       MODEL_DATA_PATH: COSMO_DATA_BASE_PATH,
       unitsConverter: gu.convertUnitsForCosmo,
       jsonRenderer: renderJSONCosmoMosmix,
@@ -229,151 +212,9 @@ function getConfigFromModel (model, COSMO_DATA_BASE_PATH, MOSMIX_DATA_BASE_PATH)
   }
 }
 
-function findClosestValidTimestamp (timestamp, modelRun) {
-  const dayDateTimeString = moment.utc(timestamp).format('YYYYMMDDTHH')
-
-  const closestModelRunTimestampSameDate =
-    moment.utc(timestamp).format('YYYYMMDD') + 'T' + modelRun
-
-  if (moment(dayDateTimeString).isAfter(moment(closestModelRunTimestampSameDate))) {
-    // Case when timestamp is not same as modelrun time and closest modelRun timestamp is an earlier date than provided
-    // E.g. 20201001T22, that means last possible modelRun for this date is not in scope
-    // So the next date's timestamp with modelRun is returned
-    const validTimestamp = parseInt(
-      moment
-        .utc(closestModelRunTimestampSameDate)
-        .add(1, 'day')
-        .format('x')
-    )
-    return validTimestamp
-  } else {
-    // Case when timestamp is not same as modelrun time and closest modelRun timestamp is returned
-    // E.g. 20201001T10 -> 20201001T21, if 21 is modelRun (by default it is)
-    const validTimestamp = parseInt(
-      moment.utc(closestModelRunTimestampSameDate).format('x')
-    )
-    return validTimestamp
-  }
-}
-
-async function loadCosmoData (
-  voiConfig,
-  referenceTimestamp,
-  lat,
-  lon,
-  gribBaseDirectory
-) {
-  let timeseriesData
-  if (voiConfig.cosmo.functionType === 'loadBaseValue') {
-    timeseriesData = await gf.loadBaseValue({
-      gribBaseDirectory,
-      referenceTimestamp,
-      voi: voiConfig.cosmo.options.key,
-      location: {
-        lat,
-        lon
-      },
-      sourceUnit: voiConfig.cosmo.options.unit,
-      targetUnit: voiConfig.target.unit
-    })
-  } else if (voiConfig.cosmo.functionType === 'loadRadiationValue') {
-    timeseriesData = await gf.loadRadiationValue({
-      gribBaseDirectory,
-      referenceTimestamp,
-      voi: voiConfig.cosmo.options.key,
-      location: {
-        lat,
-        lon
-      },
-      sourceUnit: voiConfig.cosmo.options.unit,
-      targetUnit: voiConfig.target.unit
-    })
-  } else if (voiConfig.cosmo.functionType === 'load2DVectorNorm') {
-    timeseriesData = await gf.load2DVectorNorm({
-      gribBaseDirectory,
-      referenceTimestamp,
-      voi1: voiConfig.cosmo.options.voi1_key,
-      voi2: voiConfig.cosmo.options.voi2_key,
-      location: {
-        lat,
-        lon
-      },
-      sourceUnit1: voiConfig.cosmo.options.voi1_unit,
-      sourceUnit2: voiConfig.cosmo.options.voi2_unit,
-      targetUnit: voiConfig.target.unit
-    })
-  } else if (voiConfig.cosmo.functionType === 'load2DVectorAngle') {
-    timeseriesData = await gf.load2DVectorAngle({
-      gribBaseDirectory,
-      referenceTimestamp,
-      voi1: voiConfig.cosmo.options.voi1_key,
-      voi2: voiConfig.cosmo.options.voi2_key,
-      location: {
-        lat,
-        lon
-      },
-      sourceUnit1: voiConfig.cosmo.options.voi1_unit,
-      sourceUnit2: voiConfig.cosmo.options.voi2_unit,
-      targetUnit: voiConfig.target.unit
-    })
-  }
-
-  const result = {
-    label: voiConfig.target.key,
-    unit: voiConfig.target.unit,
-    data: timeseriesData.timeseriesData,
-    location: timeseriesData.location
-  }
-  return result
-}
-
-async function readDataCosmoD2 (
-  WEATHER_DATA_BASE_PATH,
-  voisConfigs,
-  startTimestamp,
-  endTimestamp,
-  stationId,
-  modelRun,
-  stationCatalog
-) {
-  const stationObject = su.getStationById(stationCatalog, stationId)
-  const stationLocation = su.getStationLocation(stationObject)
-  const lat = stationLocation.latitude
-  const lon = stationLocation.longitude
-
-  // Works with old and new directory structure.
-  const timeseries = []
-  let closestValidTimestamp = findClosestValidTimestamp(startTimestamp, modelRun)
-  while (closestValidTimestamp <= endTimestamp) {
-    const gribBaseDirectory = getGribDirectory(
-      closestValidTimestamp,
-      WEATHER_DATA_BASE_PATH
-    )
-    for (let i = 0; i < voisConfigs.length; i += 1) {
-      try {
-        const data = await loadCosmoData(
-          voisConfigs[i],
-          closestValidTimestamp,
-          lat,
-          lon,
-          gribBaseDirectory
-        )
-        timeseries.push(data)
-      } catch (error) {
-        log.warn(error)
-      }
-    }
-    closestValidTimestamp += hoursInDay * secInHour * milliseconds
-  }
-  return timeseries
-}
-
-exports.getGribDirectory = getGribDirectory
 exports.renderJSONCosmoMosmix = renderJSONCosmoMosmix
 exports.readMosmixTimeseriesData = readMosmixTimeseriesData
 exports.renderForecastCSVCosmoMosmix = renderForecastCSVCosmoMosmix
 exports.dropTimeseriesDataNotOfInteresWithParameter = dropTimeseriesDataNotOfInteresWithParameter
 exports.checkIfValidModelRun = checkIfValidModelRun
 exports.getConfigFromModel = getConfigFromModel
-exports.readDataCosmoD2 = readDataCosmoD2
-exports.loadCosmoData = loadCosmoData
